Add draggable icon functionality
This code takes a source which would be an icon for this project and allow a user to move that icon across the screen. This function uses an observer design pattern by waiting for an event to happen to a particular source (icon) and sets it to a new location based on where the user lets go of the icon.