Added Sidebar and modified footer
When you click customize button on the footer, it now opens a sidebar. you can open and close the sidebar with the same button.

Used useState hook to determine if the sidebar was visible or not and created a toggleSidebar function that changes the state every time it is clicked